backend/pyproject.toml
[project]
name = "accounting-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.6",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "requests>=2.33.1",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.43.0",
]
